+++ DirectXFramework/DX12/RootSignature.h
@@ -124,8 +124,8 @@ namespace DX12
             {
                 CD3DX12_ROOT_PARAMETER& param;
                 CD3DX12_DESCRIPTOR_RANGE& range;
-                std::map<size_t, table_info>& tables;
-                size_t index;
+                std::map<UINT, table_info>& tables;
+				UINT index;
 
 
                 void operator=(const DescriptorTable& table)
@@ -156,7 +156,7 @@ namespace DX12
                 {
                     param.InitAsConstants(table.count, table.buffer_offset, 0, static_cast<D3D12_SHADER_VISIBILITY>(table.visibility));
                 }
-                helper(size_t index, CD3DX12_ROOT_PARAMETER& _param, CD3DX12_DESCRIPTOR_RANGE& _range, std::map<size_t, table_info>& tables): index(index), param(_param), range(_range), tables(tables)
+                helper(UINT index, CD3DX12_ROOT_PARAMETER& _param, CD3DX12_DESCRIPTOR_RANGE& _range, std::map<UINT, table_info>& tables): index(index), param(_param), range(_range), tables(tables)
                 {
                 }
 
@@ -165,7 +165,7 @@ namespace DX12
                 }
             };
 
-            helper operator[](size_t i)
+            helper operator[](UINT i)
             {
                 if (i >= parameters.size())
                     parameters.resize(i + 1);
@@ -217,7 +217,7 @@ namespace DX12
                 sampler.RegisterSpace = space;
                 sampler.ShaderRegister = i;
             }
-            std::map<size_t, table_info> tables;
+            std::map<UINT, table_info> tables;
 
         private:
             std::vector<CD3DX12_ROOT_PARAMETER> parameters;
